RAGpython/requirements.txt
pymongo
google-cloud-aiplatform
beautifulsoup4
fastapi
uvicorn
pydantic
dotenv